feat: add detail model and fix title for prompt and navigation
